refactor(agent): extract per-provider credential refresh into AgentCredentialRefreshMixin (god-file Phase 3)
AIAgent is 6,790 lines and 251 methods, the largest class in run_agent.py.
This lifts its credential-refresh cluster into run_agent_credentials.py,
following the same mechanical mixin extraction and the same naming as the
SessionDB mixins in hermes_state_*.py.

The cluster is five sibling methods with one shape: when a request is about to
go out, or has just failed on auth, re-resolve that provider's credentials and
swap them into the already-built client rather than rebuilding the world.
Codex, Nous, Vertex, Copilot and Anthropic each need a different dance to get
there, which is why it is 242 contiguous lines rather than one method with a
switch.

Moved verbatim (4988-5233):

  _try_refresh_codex_client_credentials      _try_refresh_copilot_client_credentials
  _try_refresh_nous_client_credentials       _try_refresh_anthropic_client_credentials
  _try_refresh_vertex_client_credentials

Behavior-neutral. Every body is byte-identical to what it replaced; the only
edits are the class declaration, one import, and the new module header.

Every provider dependency was already imported lazily inside the method that
uses it (from hermes_cli.auth import ..., from agent.vertex_adapter import
..., and so on). Those imports are untouched and resolve from the modules they
name, so the move does not reach them; only three module-level names had to
follow, and they come from their original sources: get_provider_request_timeout
from hermes_cli.timeouts, env_float from utils, and logger.

logger is bound by explicit name, so records keep the name run_agent and
getLogger returns run_agent's own object, verified at runtime.

run_agent.py 7,416 -> 7,170 lines; the class 6,790 -> 6,543 lines and
251 -> 246 methods.
